Move new component to a separate file
